compile eosio.system contract with multiple cpp files
