+++ contracts/eosio.system/CMakeLists.txt
@@ -1,4 +1,12 @@
-add_contract(eosio.system eosio.system ${CMAKE_CURRENT_SOURCE_DIR}/src/eosio.system.cpp)
+add_contract(eosio.system eosio.system
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/eosio.system.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/delegate_bandwidth.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/exchange_state.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/native.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/producer_pay.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/rex.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/src/voting.cpp
+)
 
 target_include_directories(eosio.system
    PUBLIC
