+++ contracts/eosio.system/include/eosio.system/exchange_state.hpp
@@ -1,9 +1,10 @@
 #pragma once
 
 #include <eosio/asset.hpp>
+#include <eosio/multi_index.hpp>
 
 namespace eosiosystem {
-   
+
    using eosio::asset;
    using eosio::symbol;
