+++ contracts/eosio.system/src/exchange_state.cpp
@@ -1,6 +1,12 @@
 #include <eosio.system/exchange_state.hpp>
 
+#include <eosio/check.hpp>
+
+#include <cmath>
+
 namespace eosiosystem {
+
+   using eosio::check;
 
    asset exchange_state::convert_to_exchange( connector& reserve, const asset& payment )
    {
@@ -24,7 +30,7 @@ namespace eosiosystem {
       const double Fi = double(1) / reserve.weight;
 
       double dR = R0 * ( std::pow(1. + dS / S0, Fi) - 1. ); // dR < 0 since dS < 0
-      if ( dR > 0 ) dR = 0; // rounding errors 
+      if ( dR > 0 ) dR = 0; // rounding errors
       reserve.balance.amount -= int64_t(-dR);
       supply                 -= tokens;
       return asset( int64_t(-dR), reserve.balance.symbol );
