+++ contracts/eosio.system/src/native.cpp
@@ -1,0 +1,27 @@
+#include <eosio.system/native.hpp>
+
+#include <eosio/check.hpp>
+
+namespace eosio {
+   bool is_feature_activated( const eosio::checksum256& feature_digest ) {
+      auto feature_digest_data = feature_digest.extract_as_byte_array();
+      return internal_use_do_not_use::is_feature_activated(
+         reinterpret_cast<const ::capi_checksum256*>( feature_digest_data.data() )
+      );
+   }
+
+   void preactivate_feature( const eosio::checksum256& feature_digest ) {
+      auto feature_digest_data = feature_digest.extract_as_byte_array();
+      internal_use_do_not_use::preactivate_feature(
+         reinterpret_cast<const ::capi_checksum256*>( feature_digest_data.data() )
+      );
+   }
+}
+
+namespace eosiosystem {
+
+   void native::onerror( ignore<uint128_t>, ignore<std::vector<char>> ) {
+      eosio::check( false, "the onerror action cannot be called directly" );
+   }
+
+}
